+++ src/account_call_builder.js
@@ -19,17 +19,6 @@ export class AccountCallBuilder extends CallBuilder {
      * Returns information and links relating to a single account.
      * The balances section in the returned JSON will also list all the trust lines this account has set up.
      *
-     * @deprecated use accountId method instead
-     */
-    address(id) {
-        console.warn("AccountCallBuilder#address is deprecated, please use AccountCallBuilder#accountId instead");
-        return this.accountId(id);
-    }
-
-    /**
-     * Returns information and links relating to a single account.
-     * The balances section in the returned JSON will also list all the trust lines this account has set up.
-     *
      * @see [Account Details](https://www.stellar.org/developers/horizon/reference/accounts-single.html)
      * @param {string} id For example: `GDGQVOKHW4VEJRU2TETD6DBRKEO5ERCNF353LW5WBFW3JJWQ2BRQ6KDD`
      * @returns {AccountCallBuilder}